re-organized application code structure
