+++ src/components/header/navigation/cart-dropdown/cart-item/cart-item-component.jsx
