+++ src/components/header/navigation/cart-dropdown/cart-item/cart-item-styles.jsx
