+++ src/components/products-preview/categories-preview/categories-preview-styles.scss
