+++ src/components/products-preview/products-preview-component.jsx
@@ -1,7 +1,7 @@
 import { Divider, Stack } from '@mui/material';
 import { useSelector } from 'react-redux';
 import { selectCategoriesMap } from '../../store/reducers/categories/categories-selector';
-import CategoryPreview from '../categories-preview/categories-preview-component';
+import CategoryPreview from './categories-preview/categories-preview-component';
 
 const ProductsPreview = () => {
   const categoriesMap = useSelector(selectCategoriesMap);
